From: James Rowe <Jay@jnrowe.uklinux.net>
I hadn't even realised until about 10 minutes ago that patch -v checks
for Sun don't output correctly(ssh session).  Minor fix for that.

A few context changes for AC_MSG_ERROR.

And a test for that buggy as hell patch 2.0 that turned up on a box a
week or so ago.  I've noted the fact it doesn't like creating
directories during patching, but it also doesn't like removing
files(which is perhaps less of a problem, it just leaves a 0 size file).