add, files, remove: accept file name with spaces
http://savannah.nongnu.org/bugs/?19477 and
http://lists.nongnu.org/archive/html/quilt-dev/2009-09/msg00010.html
reported that the current quilt does not work with file name with
spaces. we all know that it is tough to handle it with bash script and
fix all quilt commands.

this patch only fixes add, files, and remove command with normal code
path.  no excessive test has been done but a few test code is also
added.  hope this starts fixing quilt to handle all possible path
pattern.