Remember QUILT_PATCHES and QUILT_SERIES by storing them in .pc
When the quilt db is created/initialized, QUILT_PATCHES and QUILT_SERIES
are written down in .pc/.quilt_{patches,series} and in subsequent
invocations of quilt those values are reused if the environmement
variables are no longer set.

When identifying the root directory, quilt now also looks for the
existence of .pc and not only of QUILT_PATCHES since that value might not
be set and we must be able to look it up from .pc/.quilt_patches instead.

Based on a preliminary patch made by Goswin von Brederlow
<goswin-v-b@web.de> (see http://bugs.debian.org/557623).