Separate working directory for each test case
Let the test suite runner script create a separate working
directory for each test. This makes it possible to run several
tests in parallel, speeding up testing on SMP systems. It also
ensures that a broken test case can no longer cause another test
case to fail.

At this point, it is possible to run the test suite in parallel,
the final result (success or failure) will be correct, however
the output is hardly readable because the progress of all tests
are interlaced. This will be improved later.

As a nice side effect, this means we can drop the setup and
clean-up steps that were present in all test cases.