diff, refresh: Accept file names with spaces
This is a continuation of Yasushi SHOJI's initial work of fixing
support for file names with spaces. This time I fixed support in the
diff and refresh commands, two commands which are frequently used.
As before, no guarantee is made that all options are covered, but
this should be a good starting point.

Signed-off-by: Jean Delvare <jdelvare@suse.de>