configure.ac: Fix test for GNU patch version
We currently ask for GNU patch >= 2.4, but the test suite doesn't
actually pass with GNU patch 2.4. It passes with GNU patch 2.5 with
minor edits to be more tolerant to the exact output of "patch".

I have no idea how much work it would be to get 2.4 to be supported
again, but it doesn't seem unreasonable to ask for 2.5 which was
released in August 1997. So let's just do that for now. If anyone
badly misses support for GNU patch 2.4, well, we accept patches ;)

Signed-off-by: Jean Delvare <jdelvare@suse.de>