Don't let TMPDIR take over /var/tmp
TMPDIR is set to /tmp by default on some systems [1], in this case
letting it take over /var/tmp voids the point of using /var/tmp for
data size reasons in the first place [2]. I'd rather hardcode
/var/tmp. But in case someone really wants to control this, we can
check if VARTMPDIR is set and use that if it is set.

[1] I tried to track down how it was happening and I am far from
certain that this is on purpose. I've seen it happen only on local
graphical login, not on text console login nor remote login. This
suggests a leak from some piece of the graphical desktop stack. And
I've seen it on both KDE and Gnome, but not Xfce. Go figure.

[2] This is the reason given in commit
d4074d26276b9a2d452563dbf57ed497afa7b710 for using /var/tmp.