Revert "Fix auto-generation of quilt.1 to work with dash"
This reverts commit 5f17e6f29c22282a4bf2819ac7ee08c942e6e103.

We rely on bash for the rest of quilt, so it is no real issue to rely
on it for the makefile. If we try to remove it, we have to deal with
the portability another way. For example, BSD and Mac OSX versions of
sed are not able to deal with \n by themselves...

The correct fix will be to force the SHELL to be bash in the makefile.