Silence the completion of the push script
When using bash TAB completion in "quilt push <TAB>", if no patch is
applied, it shows the standard error of the command "quilt applied"
("No patches applied"), used in the source file bash_completion to get
the list of applied patches.

This is similar to commit 6af132b5061f6773f9591ebde625c92a395ebc91,
but for the push script that were forgotten previously.