backup-files: Try mass copy first on copy
When copying many files to a snapshot directory, try a mass copy
first, as it is much faster. It is however not portable and may thus
fail. If it fails, fallback to per-file processing, which always
works.

This change results in a huge performance boost on systems where
the cp command supports all the required options (which includes
all systems using GNU coreutils.)

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Reviewed-by: Raphael Hertzog <hertzog@debian.org>