pop: Fix timestamp comparison
The timestamps recorded on file systems have a limited resolution.
For this reason, a patch or file with a timestamp exactly equal to
.timestamp should be considered as more recent than .timestamp, i.e.
potentially modified since the last push or refresh. This is
particularly important for the test suite, as we can do many actions
in the same second there.