pop: Simplify check_for_pending_changes
We have to patch temporary files when checking if a patch can be
popped without losing pending changes. When this fails, the status
code we get from "patch" isn't accurate enough to draw conclusions
so don't even try. Just compare the result with the current working
files and tell the user if there is any difference or not.