test: ignore leading whitespace in wc output.
POSIX-compliant implementations of wc right-justify each column of
output in 7 character wide columns.  Tests that verify the output
of wc -l should take into account the number of digits in the line
count and add the appropriate padding.  Therefore, in order for the
tests to pass with both POSIX-compliant and non-POSIX-compliant
implementations of wc tests should match wc output against patterns
with optional whitespace prefixes.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>