test/run: Fix variable substitution
Variable substitution on command lines was happening too early so
setting a variable only took effect with the second next line of the
test case.

Additionally, there was no check that the environment variable was
actually set. This resulted in perl warnings during the test suite if
a variable is ever used before having been set.

Fix both issues.

Also add a test case for this feature, to avoid a future breakage.
Some of the tester script features are tricky and easy to get wrong,
so test them independently of quilt in a dedicated test case.