Stop using patch option -E
GNU patch version 2.5 and newer deals just fine with empty files,
however our use of option -E breaks this feature. So stop using this
option. Anyway, the manual page says: "Normally this option is
unnecessary" and the test suite passes just fine without it.

If anyone is using legacy patch formats which do require the use of
option -E, they can add it to QUILT_PATCH_OPTS to revert to the
previous behavior.