Fix refreshing of patches emptying a file
Quilt can now import patches emptying a file, however the proper
header formatting is lost on refresh. Fix it.

The format is still not correct for patches creating a file. This is
because quilt creates an empty backup file for every non-existent file
being added to a patch. This is how quilt remembers that the file is
modified by the top-most patch. Ideally the list of files belonging to
the top-most patch should be stored separately in a metadata file.
That being said, it is not critical in practice because GNU patch
makes no difference between adding contents to a new file and adding
contents to an empty file.