patchfns: Simplify function find_patch_file further
find_patch_file is only used by "quilt import". It is called on patch
file names passed on the command line. The first part of the function
clearly looks for "external" patch files (with absolute or relative
paths.)

The second part looks for the patch in the series file. There are 3
possibilities:
* If the patch is not in series, then the import fails with:
Patch foo.patch does not exist
* If the patch is in series, applied, then the import fails with:
Patch foo.patch is applied
* If the patch is in series, not applied, then the import fails with:
Patch foo.patch already exists in series.

So this part of the function makes no sense and should be deleted.