diff: Workaround bash bug
Old versions of bash (at least version 3.2.51) don't properly handle
prefix stripping together with quoting when evaluating an array. So
strip the prefix before adding each file to opt_files. It's faster
anyway.

Same thing when diffing against a snapshot, strip the snapshot
directory prefix from file names before evaluating the quoted files
array.

This fixes a regression introduced in:

commit b0baeeb6b61132af92fd75df5f912554d295dee1
Author: Jean Delvare <jdelvare@suse.de>
Date:   Fri Mar 25 18:48:49 2011 +0100

    diff, refresh: Accept file names with spaces

(Only affecting the versions of bash which have the aforementioned bug.)

This also fixes a bug when called from a subdirectory and a file
passed as an argument starts with "./".

Extend the test suite to test both cases, so that such bugs can't
sneak in in the future.