Let print_patch print a newline character
Function print_patch currently doesn't print a newline character at
the end of the patch name. Because of this, callers which need the
newline character much wrap print_patch with an echo. This hurts
performance and isn't too elegant either.

So let print_patch print a trailing newline character. If the caller
doesn't need it, it will typically be stripped by bash anyway (tailing
whitespace is stripped on variable expansion.) And if the caller does
need it, it avoids the extra call to echo.