Delete quilt.changes
File quilt.changes is essentially redundant with git log. It is close
to useless to packagers and users as it contains too much details
while at the same time some important changes can be missing when we
forget to include them in the file.

A new file, NEWS, was created to list the user-visible change summary
for each version. This is much more useful for the target audience.
NEWS is now used instead of quilt.changes to generate the %changelog
section of the rpm spec file.

Thus we can delete quilt.changes, as well as changes2changelog.