inspect: Handle long options passed to tar
The command line interface to tar is complex and sometimes confusing,
but we should still do our best to figure out where the file name is
on that command line.

Add support for the --file FILE and --file=FILE options. Other long
options must be explicitly skipped, as well as short options not
containing the letter "f".

With this we should be good to go in most real-world cases, but
there are still a few corner cases we may not handle properly. Let's
just hope we never hit them.

Reported by Petr Tesarik.