Check for series file consistency
Quilt allows manual changes to the series file to some degree. For
example, adding comments or reordering patches in the unapplied
section of the series file is OK. However, changing the order of
applied patches breaks a number of assumptions and can cause quilt to
produce unexpected or confusing results.

For example, starting from this:

+ patches/01.patch
= patches/02.patch
  patches/03.patch
  patches/04.patch

and moving the last patch at the beginning of the series file,
"quilt series -v" will print:

+ patches/04.patch
+ patches/01.patch
= patches/02.patch
  patches/03.patch

That is, it will claim that 04.patch is applied, while it it not.
Likewise, 04.patch would be listed by neither "quilt applied" nor
"quilt unapplied".

While addressing all such cases would certainly be possible, that
would require a significant amount of work, and would come with
performance penalties. It would also be difficult to be certain that
all issues have been found and addressed. So it seems more reasonable
to simply spot such manual changes to the series file and ask the user
to pop all patches to start from a clean state as needed.