setup: Add --fast option
This is an alternative implementation of "quilt setup" for rpm spec
files, which is much faster than the original implementation. The
idea is to have rpmbuild generate our working tree directly, instead
of taking note of everything it does and then replaying that record
from scratch.

The new implementation is enabled with option --fast. The original
implementation can be selected with --slow, which is the default.
Having this option makes it possible to set the default to --fast in
~/.quiltrc and revert to --slow on the command line on a case-by-case
basis. This will also make it easier if we decide to change the
default in the future.

In general, the generated working tree should be the same with --slow
or --fast.  There are 3 known exceptions though:
* The header of the series files is incomplete when using --fast,
  which means that you can't reuse these series files to feed "quilt
  setup" later.
* If the spec file generates files other than with the tar, unzip and
  patch commands, you wouldn't get these files in your working tree
  with --slow, but with --fast you will.
* With --fast, all patches are added to the series file, regardless
  of whether they apply or not. This also means that patches which
  fail to apply are not reported as such until you attempt to push
  them.

On large packages, this alternative implementation was found to be
up to 4 times faster than the original implementation.