mail: Don't include the release number in User-Agent
@RELEASE@ is replaced with the package-level release number in two
places: the spec file, where it is legitimate, and the mail command's
User-Agent string, where I think it is not. The package release
number should not affect the contents of the package, otherwise it
makes it difficult to compare build results.

I have checked other MUA (Mutt, Thunderbird, Claws Mail) and they all
only include the version, not the release number in their User-Agent
string.

Drop the @RELEASE@ replacement rule to make sure we don't use it
accidentally anywhere in the future. The spec file has its own build
rule so it is not affected.