edit: Use command "remove" to remove files
Commit f1c186ee renamed command "remove" to "revert". Commit bd1dfdab
changed the semantics of command "revert". Commit 6d2501ac restored
command "remove". However command "edit" is still calling command
"revert" to remove files from the patch when it should be using
command "remove".

This case was not covered by the test suite, which is why the bug was
not spotted earlier.