quilt-el: fix quilt-editable when patches are stored in subdirs
I frequently use sub-directories to store different patchsets under
the quilt patches/ directory. Unfortunately, the quilt emacs mode
does not handle this case and the file is not considered editable.

The issue seems to be in the quilt-top-patch which does an equivalent
of basename on the patch name. The proposal below changes the
quilt-editable routine to keep the full name of the patch.

Signed-off-by: Cédric Le Goater <legoater@free.fr>