bash_completion: Drop duplicate functions
Completing long options of the grep command is not a fundamental
feature. I can't remember ever needing it. The most useful grep
options are short ones by design. Additionally, our internal
implementation of _longopt is broken as far as I can tell.

So drop our implementations of _expand, _filedir and _longopt, and
only call _longopt if it is already available from bash-completion.