bash_completion: Delete dead code
You can't change the caller's environment from the right hand side of
a pipe. So that code was doing nothing, which means we can just
delete it.

What the code was supposedly doing, is actually implemented by
complete's option "filenames". So move the comment to where this
option is set, to clear the confusion.