patch-wrapper: Make the patch names that quilt shows more convenient
When operating on an rpm package, create links to the rpm sources and
build directories and make patch names relative to these links. This is
more convenient than using ugly absolute paths.

This patch has been in the SUSE quilt package since February 2006, it
seems to be about time to get it upstream.