test: Consistently use %{P} and %{_P}
There is no point in defining P and _P in the Makefile to allow for
QUILT_PATCHES_PREFIX to be set or unset if we do not use these
variables consistently in the test suite. Replace hardcoded
"patches/" with %{P} or %{_P} throughout the test suite wherever
relevant so that unsetting QUILT_PATCHES_PREFIX in test.quiltrc
actually works.

Signed-off-by: Jean Delvare <jdelvare@suse.de>