mail: Remove Content-Disposition headers
The Content-Disposition headers quilt has been adding since 2015 are
causing problems with Thunderbird and Gmail, which display patches as
attachments instead of text if this header is present and includes a
filename.  This breaks email-based patch reviewing and isn't likely to
get fixed on the email client side, so stop adding those headers.

Recipients of patches via email already need a mechanism to construct
filenames for messages coming from anything but quilt, so this change is
unlikely to cause major problems for existing users.