compat/getopt: Allow non-digit parameter embedded in short option
The compatibility getopt script allows only digit parameters to be
embedded in short options. Util-linux's getopt implementation does
not have such a restriction and allows any parameter to be embedded
in short options. As a consequence, using the compatibility getopt
script would choke for example on "-pab", which is a legal option
of the "quilt refresh" command.

Remove the limitation on digits so that the compatibility getopt
script allows what util-linux allows. This fixes the second half
of bug #54772:
https://savannah.nongnu.org/bugs/index.php?54772

As a side note, this feature of the compatibility script was broken
anyway, as it would output the digits in reverse order.

Signed-off-by: Jean Delvare <jdelvare@suse.de>