compat/getopt: Handle a second separator
getopt can be passed 2 '--' separators. The first one tells that
getopt options are over and target program options start. The second
one tells that the target program's options are over and following
arguments should be treated as non-options even if they look like
options.

This second separator was not handled, causing the compatibility
getopt script to treat the following arguments as options, eventually
failing one way or another.

Properly detect and handle the second separator. This fixes the first
half of bug #54772:
https://savannah.nongnu.org/bugs/index.php?54772

Signed-off-by: Jean Delvare <jdelvare@suse.de>