new: Fix series file corruption when trailing newline is missing
When editing the series file with certain editors, it is possible that
no trailing newline character is present on the last line. In such
case, adding a new patch at the end of the series would corrupt the
series file. Prevent that by ensuring that a newline character is
always present.

This fixes the second issue reported in bug #54295:
https://savannah.nongnu.org/bugs/?54295

Signed-off-by: Jean Delvare <jdelvare@suse.de>