test: Don't hard-code .pc
Use a variable (QUILT_PC) instead of hard-coding ".pc" as the name of
the directory where quilt stores the state of the working tree
throughout the test suite. We still set that variable to ".pc" to
make it easier to investigate failed test cases, but at least this
allows us to promptly test that QUILT_PC can actually be set to any
other value and quilt still works.

Signed-off-by: Jean Delvare <jdelvare@suse.de>