setup: Don't obey the settings of any englobing .pc
This is mainly intended to get the setup.test working even if the
debian package contains a .pc directory. Without this patch, the
debian packaging stuff will get the testsuite using debian/patches
instead of patches (because it's the way it goes in our .pc). The
test breaks with that setting.

The patch changes the setup command to not take the settings of any
.pc directory found, and reset QUILT_PC QUILT_PATCHES and QUILT_SERIES
to their default values.

This fixes Debian bug #573689:
http://bugs.debian.org/573689