Call pager with original LANG environment variable
Author: Urs Thuermann <urs@isnogud.escape.de>
Date:   Tue Feb 11 06:43:05 2020 +0100

Call pager with original LANG environment variable

The first few lines in quilt/scripts/patchfns set the LANG environment
variable to POSIX saving the previous value in ORIGINAL_LANG.  This is
used to restore the original locale when calling a user-specified diff
command but not for the pager.  Without this patch the pager is called
with LANG=POSIX causing non-ASCII characters to be printed wrong
unless LC_CTYPE is also set in the environment.

This patch restores LANG for the pager to its original value.