setup: Default to fast mode
The fast mode of processing spec files has been implemented over 4
years ago. Not only it is faster, but it also handles corner cases
in spec files better in general. Make it the new default.

Option --slow is still available to revert to the old implementation
if needed. However in the long run I hope that we can get rid of it
completely.

Signed-off-by: Jean Delvare <jdelvare@suse.de>