Consistently complain early if no series file is found
If no series file is found, let all quilt commands which need it
complain about it immediately. This aligns the behavior of the "pop",
"series" and "snapshot" commands with how all other commands behave.

The "import", "new", "setup" and "upgrade" commands are not affected
by this change, as they are legitimately called without a series
file.

A side effect of this change is that the "pop", "push", "top" and
"next" commands will now return with error code 1 instead of 2 when
called outside of a quilt-managed tree.

Inspired by Martin Quinson. This fixes Debian bug #369908:
https://bugs.debian.org/369908

Signed-off-by: Jean Delvare <jdelvare@suse.de>