quilt.el: Make quilt-patches-directory return per-project setting
When creating the .pc/ directory, quilt writes the value of
QUILT_PATCHES to .pc/.quilt_patches. On all subsequent invocations,
quilt uses the contents of .pc/.quilt_patches as the value of
QUILT_PATCHES, rather than the value set in quiltrc.

All the callers in quilt.el really expect the per-project setting from
.pc/.quilt_patches, so return the value set therein if present.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>
Signed-off-by: Jean Delvare <jdelvare@suse.de>