Quote parameter to gen_tempfile as needed
When the parameter passed to gen_tempfile is based on the working
directory, we need to quote it because it could contain spaces or
other special characters. Also quote the string returned by this
function for the same reason.

Affected commands:
* quilt diff -z
* quilt fold
* quilt refresh -z
* quilt revert

Also test these code paths in the test suite to avoid regressions.

Signed-off-by: Jean Delvare <jdelvare@suse.de>