inspect-wrapper: procfs resolves links
When patch files are passed through stdin, we get the actual patch
file name from procfs. It turns out that procfs resolves symbolic
links, and that breaks our later attempt to strip the prefix from
the path to extract a relative path to the patch file.

This is solved by also resolving symbolic links in the prefix
before stripping it.

Signed-off-by: Jean Delvare <jdelvare@suse.de>