Tighten the patch format parsing
Require a "---" line to be followed by a "+++" line to recognize it as
the start of a unified diff.  Likewise, require a "***" line to be
followed by a "---" line to recognize it as the start of a context diff.

Without that, a line like "*** NOTE ***" in the patch header will cause
the refresh and header commands to truncate the header.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>