Delay sourcing patchfns until options have been processed
Commit 8b39a960afcf ("Consistently complain early if no series file
is found") unexpectedly broke commands like "quilt add -h", because
the check for a series file happens before the command line options
are processed. This, in turn, breaks the generation of the quilt
manual page, which relies on the output of each command when called
with option -h.

Reorder the code so that options are always processed first, and then
patchfns is sourced. That way, option "-h" will work again even if
not in a quilt working tree.

Fixes: 8b39a960afcf ("Consistently complain early if no series file is found")
Signed-off-by: Jean Delvare <jdelvare@suse.de>