Move the code which adds the default parameters from quiltrc
Move the code which reads the default parameters for each command
from quiltrc, from patchfns to the main quilt script. This is needed
because the parsing of command line options needs to happen before we
source patchfns, which is obviously not possible if some of the
options are being set by patchfns itself.

Signed-off-by: Jean Delvare <jdelvare@suse.de>