import: Fix option -P
I broke "import -P" with my previous commit. Using this option now
fails with the following error message:
/usr/share/quilt/import: line 168: [: missing `]'

This is caused by a missing space before the closing square bracket,
which prevents bash from parsing the line.

Fix that, and add a test case to cover this code path, so that no
such bug can creep in in the future.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Fixes: c539338458e4 ("Delay sourcing patchfns until options have been processed")