Reuse the shell
One side effect of commit ca85fbd82022 ("Move the code which adds the
default parameters from quiltrc") is that we now source the quiltrc
configuration file twice. While parsing this file shouldn't take
overly long, this still feels conceptually wrong.

This can be avoided by reusing the same shell for the quilt command
as we used for the quilt "launcher" itself, instead of starting a
new instance of bash for it.

As a nice side bonus, this makes quilt about 2% faster (measured on
the test suite).

Signed-off-by: Jean Delvare <jdelvare@suse.de>