Man page: escape ASCII hyphen-minus characters used as such
When a "literal" ASCII 45 hyphen-minus character is desired, it has to
be escaped; this is a requirement going all the way back to Bell Labs
Troff, but is frequently overlooked.  Getting it right enables
accurate cut-and-paste of code examples, filenames, URLs, and so forth
from roff documents prepared for UTF-8 terminal and PDF output, possibly
among others.

See section 2.1 of CSTR #54, "Troff User's Manual", Ossanna & Kernighan
<https://www.troff.org/54.pdf>.