Man page: rewrite discussion of QUILT_COLORS configuration variable
These are ANSI escape sequences as defined by ECMA-48; recast the entire
discussion in light of that fact.

Condense the many tagged paragraphs with a templated discussion of
defaults into a table.

Sort the QUILT_COLORS format names into alphabetical order.

Add a hint to the formatter (man) to call the tbl preprocessor to
render the table.

Expand the example to be more demonstrative.

Add pointers to the ECMA-48 standard document and the console_codes
section 4 man page (from Michael Kerrisk's man-pages project, widely
available) to the See Also section.

[JD: Some fixes, added patch_offs to the table, added a note that the
     patches command shares colors with the series command]