setup: Document the limitation of spec file support
Quilt setup can't be used on arbitrary spec files. Due to the fact
that quilt only deals with patches, every other modification to the
working tree has to happen first, and the patches must be applied
last.

This is a design limitation and not a bug, but document it so that
the users are aware of it and can adjust the %prep section of their
spec file if needed.

Signed-off-by: Jean Delvare <jdelvare@suse.de>