test: Fix a race condition
The test suite does not differentiate between stdout and stderr. When
messages are printed to both, the order in which they will reach us
is apparently not guaranteed. Ideally this would be deterministic, but
until then, explicitly test stdout and stderr separately in the test
case itself. Otherwise the test suite fails randomly, which is a pain
for distribution package maintainers.

This fixes bug #63651 reported by Ross Burton:
https://savannah.nongnu.org/bugs/index.php?63651

Signed-off-by: Jean Delvare <jdelvare@suse.de>