config: Update install-sh to version 2016-01-11.22
The install-sh script we ship has issues. It is not able to install
multiple files at once, and is unable to create directories when all
their parent elements do not already exist.

Update to a more recent version of that script where these issues are
fixed. Also set the executable bit so that the script can actually be
run.

This fixes bug #64430:
https://savannah.nongnu.org/bugs/index.php?64430

Signed-off-by: Jean Delvare <jdelvare@suse.de>